moved connect to home page
